(MODULES-7174) Ensure Absent of IIS Feature Fails
Added update method to fix parent provider flush

Previously when setting the `ensure` property for an IIS feature to
`absent` the Puppet run would fail because the provider would attempt to
call the `update` method which did not exist. This change implements a
stub method so that the run can continue.